Expect data polls from a sleepy child being restored (#1315)
This commit relaxes the check in `MleRouter` for `GetNeighbor()`
and `RemoveNeighbor()` methods to consider a child in restored
or restoring (`kStateChildUpdateRequest`) states in addition to
valid state.

This change ensures that an outbound message for a sleepy child being
restored in queued until a data poll is received from the child. To
ensure that a received message from such a child is not dropped (due
to security frame counter mismatch) this commit also updates the
`Mac::ReceiveDoneTask()` to ensure that the neighbor corresponding
to message source is indeed in `kStateValid` state.