Ncp: Simplify Tx Buffer Management (#1401)
This commit changes the `NcpBase` class so that it defines and
controls the NCP Tx buffer used for outbound spinel message (instead
of delegating this to the sub-classes `NcpUart` and `NcpSpi`).

This removes the need to use abstract `virtual` methods in the base
class. It also simplifies the code by removing the common/duplicate
implementation in each sub-class.